[my::array] Constructor with variadic template
Got to know to template parameter pack and its syntax.

- c++ variadic parameters works only with _templates_
(not regular functions, that have only regular parameters)

- you can create template member function inside template class.
It can be even constructor.
And the function's template arguments can be a far cry from
the class' template argument.

- you can expand parameter pack through the recursion. This method
pointed me to my games with Erlang

TODO:
So, I've created array's constructor that accept variable number of
parameters. There's no such constructor in std::array.
It would be great to play with std::initializer_list later on.